Change the port used as an example in the Docker composition
Since no actual networking is going on inside the image this change is
non-destructive. I am making it because we are now using the
GitHubSecurityLab/actions-permissions/monitor Action and the proxy it
configures uses port 8080. There are currently no options for
configuring this in the Action and the easiest path forward in this
case is to simply change the port number used in the Docker
composition.